11-Medicine-Scanner/requirements.txt
Pillow>=10.0.0
requests>=2.31.0
numpy
